move notebook model reducer into separate file
